+++ src/chains/Sepolia/Common.sol
@@ -5,6 +5,8 @@ import {SettlerBase} from "../../SettlerBase.sol";
 
 import {IERC20} from "@forge-std/interfaces/IERC20.sol";
 import {MaverickV2, IMaverickV2Pool} from "../../core/MaverickV2.sol";
+import {UniswapV4} from "../../core/UniswapV4.sol";
+import {IPoolManager} from "../../core/UniswapV4Types.sol";
 import {FreeMemory} from "../../utils/FreeMemory.sol";
 
 import {ISettlerActions} from "../../ISettlerActions.sol";
@@ -18,10 +20,12 @@ import {
     IUniswapV3Callback
 } from "../../core/univ3forks/UniswapV3.sol";
 
+import {SEPOLIA_POOL_MANAGER} from "../../core/UniswapV4Addresses.sol";
+
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../../SettlerAbstract.sol";
 
-abstract contract SepoliaMixin is FreeMemory, SettlerBase, MaverickV2 {
+abstract contract SepoliaMixin is FreeMemory, SettlerBase, MaverickV2, UniswapV4 {
     constructor() {
         assert(block.chainid == 11155111 || block.chainid == 31337);
     }
@@ -35,6 +39,19 @@ abstract contract SepoliaMixin is FreeMemory, SettlerBase, MaverickV2 {
     {
         if (super._dispatch(i, action, data)) {
             return true;
+        } else if (action == uint32(ISettlerActions.UNISWAPV4.selector)) {
+            (
+                address recipient,
+                IERC20 sellToken,
+                uint256 bps,
+                bool feeOnTransfer,
+                uint256 hashMul,
+                uint256 hashMod,
+                bytes memory fills,
+                uint256 amountOutMin
+            ) = abi.decode(data, (address, IERC20, uint256, bool, uint256, uint256, bytes, uint256));
+
+            sellToUniswapV4(recipient, sellToken, bps, feeOnTransfer, hashMul, hashMod, fills, amountOutMin);
         } else if (action == uint32(ISettlerActions.MAVERICKV2.selector)) {
             (
                 address recipient,
@@ -65,5 +82,9 @@ abstract contract SepoliaMixin is FreeMemory, SettlerBase, MaverickV2 {
         } else {
             revert UnknownForkId(forkId);
         }
+    }
+
+    function _POOL_MANAGER() internal pure override returns (IPoolManager) {
+        return SEPOLIA_POOL_MANAGER;
     }
 }